requirements.txt
addict
tqdm
torch
numpy
scipy
addict
open3d
opencv-python
opencv_contrib_python
lxml
exifread
rawpy
Pillow
pyvista
termcolor
pylatexenc
lpips
pycolmap
scikit-image
imageio
tk
pyhocon
## Apriltags
# compile the apriltag library and python wrapper directly from https://github.com/AprilRobotics/apriltag.git (e5e234125b77b2a4ea9441d10689acce123acf3d)